some cleaning up, added readme
