+++ package.json
@@ -1,14 +1,15 @@
 {
   "name": "serverless-api-gateway-caching",
-  "version": "1.0.0-rc7",
+  "version": "1.0.0-rc8",
   "description": "",
   "main": "src/apiGatewayCachingPlugin.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "",
+  "author": "Diana Ionita",
   "license": "ISC",
   "dependencies": {
-    "aws-sdk": "^2.310.0"
+    "aws-sdk": "^2.310.0",
+    "lodash.isempty": "^4.4.0"
   }
 }
